add claim function to cgtc contract and ui updates
